+++ applications/fishing-map/src/features/timebar/Timebar.tsx
@@ -24,6 +24,7 @@ import useViewport from 'features/map/map-viewport.hooks'
 import { selectActivityCategory, selectTimebarGraph } from 'features/app/app.selectors'
 import { getEventLabel } from 'utils/analytics'
 import { upperFirst } from 'utils/info'
+import { selectIsMapDrawing } from 'features/map/map.selectors'
 import {
   setHighlightedTime,
   disableHighlightedTime,
@@ -67,6 +68,7 @@ const TimebarWrapper = () => {
   const tracks = useSelector(selectTracksData)
   const tracksGraphs = useSelector(selectTracksGraphs)
   const tracksEvents = useSelector(selectEventsWithRenderingInfo)
+  const isMapDrawing = useSelector(selectIsMapDrawing)
 
   const dispatch = useDispatch()
 
@@ -195,7 +197,7 @@ const TimebarWrapper = () => {
       : null
   }, [timebarVisualisation, showGraph, tracksGraphs])
 
-  if (!start || !end) return null
+  if (!start || !end || isMapDrawing) return null
 
   return (
     <div>