+++ applications/vessel-history/src/features/settings/components/PortVisits.tsx
@@ -1,13 +1,14 @@
 import React, { useMemo } from 'react'
 import cx from 'classnames'
-import { useTranslation } from 'react-i18next'
-import { IconButton, InputText, MultiSelect } from '@globalfishingwatch/ui-components'
+import { Trans, useTranslation } from 'react-i18next'
+import { InputText, MultiSelect } from '@globalfishingwatch/ui-components'
 import {
   PORTVISIT_EVENTS_MAX_DISTANCE,
   PORTVISIT_EVENTS_MAX_DURATION,
   PORTVISIT_EVENTS_MIN_DISTANCE,
   PORTVISIT_EVENTS_MIN_DURATION,
 } from 'data/constants'
+import DataAndTerminology from 'features/data-and-terminology/DataAndTerminology'
 import { SettingEventSectionName, SettingsPortVisits } from '../settings.slice'
 import { useSettingsConnect, useSettingsRegionsConnect } from '../settings.hooks'
 import styles from './SettingsComponents.module.css'
@@ -32,6 +33,7 @@ const PortVisits: React.FC<SettingsProps> = (props): React.ReactElement => {
       ),
     [COUNTRIES, settings.flags, getOptions]
   )
+  const eventType = useMemo(() => t(`settings.${section}.title`), [section, t])
 
   return (
     <div className={styles.settingsFieldsContainer}>
@@ -39,7 +41,30 @@ const PortVisits: React.FC<SettingsProps> = (props): React.ReactElement => {
         <div className={styles.settingsField}>
           <label>
             {t('settings.portVisits.inThesePortStates', 'Events in these port States')}
-            <IconButton type="default" size="tiny" icon="info"></IconButton>
+            <DataAndTerminology
+              size="tiny"
+              type="default"
+              title={t('settings.portVisits.inThesePortStates', 'Events in these port States')}
+            >
+              <Trans
+                i18nKey="settings.portVisits.inThesePortStatesDescription"
+                values={{ eventType }}
+              >
+                Highlight all {{ eventType }} that may have occured in a port State. A port visit
+                occurs when a vessel transits to an anchorage, and the port State is assigned based
+                on the Exclusive Economic Zone (EEZ) the anchorage is within. Anchorages are based
+                upon the Global Fishing Watch anchorages dataset, a global database of anchorage
+                locations where vessels congregate:
+                <a
+                  href="https://globalfishingwatch.org/datasets-and-code-anchorages/"
+                  rel="noopener noreferrer"
+                  target="_blank"
+                >
+                  https://globalfishingwatch.org/datasets-and-code-anchorages/
+                </a>{' '}
+                . Please investigate the highlighted events further on the map.
+              </Trans>
+            </DataAndTerminology>
           </label>
           <MultiSelect
             selectedOptions={flags.selected}
@@ -53,7 +78,17 @@ const PortVisits: React.FC<SettingsProps> = (props): React.ReactElement => {
         <div className={cx(styles.settingsField, styles.inlineRow)}>
           <label>
             {t('settings.duration', 'DURATION')}
-            <IconButton type="default" size="tiny" icon="info"></IconButton>
+            <DataAndTerminology
+              size="tiny"
+              type="default"
+              title={t('settings.duration', 'Duration')}
+            >
+              <Trans i18nKey="settings.durationDescription" values={{ eventType }}>
+                Highlight all {{ eventType }} that had a duration longer than the value configured.
+                Example if you configure 5 hours, you will see all {{ eventType }} with duration
+                more or equal to 5 hours
+              </Trans>
+            </DataAndTerminology>
           </label>
           <span>{t('settings.longerThan', 'Longer than')}</span>
           <InputText
@@ -70,7 +105,18 @@ const PortVisits: React.FC<SettingsProps> = (props): React.ReactElement => {
         <div className={cx(styles.settingsField, styles.inlineRow)}>
           <label>
             {t('event.distanceShore', 'Distance from shore')}
-            <IconButton type="default" size="tiny" icon="info"></IconButton>
+            <DataAndTerminology
+              size="tiny"
+              type="default"
+              title={t('settings.distanceShore', 'Distance from shore')}
+            >
+              <Trans i18nKey="settings.distanceShoreDescription" values={{ eventType }}>
+                Highlight all {{ eventType }} that had a distance longer than the value configured.
+                Example if you configure 8 km, you will see all {{ eventType }} with distance more
+                or equal to 8 km from the shore. Distance from shore will be highlighted if either
+                the start or end position is more or equal to the specified distance.
+              </Trans>
+            </DataAndTerminology>
           </label>
           <span>{t('settings.longerThan', 'Longer than')}</span>
           <InputText